restore timeout_virsh to virsh and adjust async secgroups

+++ scripts/kvm/async_job/clear_sg_chain.sh
@@ -3,18 +3,21 @@
 cd `dirname $0`
 source ../../cloudrc
 
-[ $# -lt 1 ] && echo "$0 <interface>" && exit -1
+[ $# -lt 1 ] && echo "$0 <interface> [force]" && exit -1
 
 vnic=$1
+force=$2
 chain_in=secgroup-in-$vnic
 chain_out=secgroup-out-$vnic
 chain_as=secgroup-as-$vnic
 
-for i in {1..35}; do
-    [ -f $async_job_dir/$vnic ] && break
-    sleep 1
-done
-rm -f $async_job_dir/$vnic
+if [ "$force" != "true" ]; then 
+    for i in {1..35}; do
+        [ -f $async_job_dir/$vnic ] && break
+        sleep 1
+    done
+    rm -f $async_job_dir/$vnic
+fi
 
 apply_fw -D FORWARD -m physdev --physdev-out $vnic --physdev-is-bridged -j secgroup-chain
 apply_fw -D FORWARD -m physdev --physdev-in $vnic --physdev-is-bridged -j secgroup-chain
